+++ podcasts/DeveloperMenu.swift
@@ -26,6 +26,7 @@ struct DeveloperMenu: View {
                 }
 
                 Button("Force Reload Discover") {
+                    DiscoverServerHandler.shared.discoveryCache.removeAllCachedResponses()
                     NotificationCenter.postOnMainThread(notification: Constants.Notifications.chartRegionChanged)
                 }